ALl updated.

+++ app/src/main/java/edu/pract5/apirestfree/data/MotorcyclesRepository.kt
@@ -21,54 +21,44 @@ class MotorcyclesRepository (
      *
      * @return Cold flow list of motorcycles.
      */
-    suspend fun getRemoteMotorcyclesByMakeOrModel(model: String): Flow<List<Motorcycle>> {
-        return remoteDataSource.getRemoteMotorcyclesByMakeOrModel(model)
-    }
-
-    suspend fun getRemoteMotorcycles(): Flow<List<Motorcycle>> {
+    fun getRemoteMotorcycles(): Flow<List<Motorcycle>> {
         return remoteDataSource.getRemoteMotorcycles()
     }
 
-    fun getLocalMotorcycles(): Flow<List<Motorcycle>> {
+    /**
+     * Get the complete list of deleted motorcycles from the local DB.
+     *
+     * @return Cold flow list of deleted motorcycles.
+     */
+    /*fun getLocalMotorcycles(): Flow<List<Motorcycle>> {
         return localDataSource.getLocalMotorcycles()
-    }
-
-    /*fun getLocalMotorcyclesByModel(model: String): Flow<List<Motorcycle>> {
-        return localDataSource.getLocalMotorcyclesByModel(model)
     }*/
 
     /**
-     * Get the complete sorted list of favourite motorcycles from the local DB.
+     * Get the complete sorted list of deleted motorcycles from the local DB.
      *
      * @param filter Ascendant or descendant sorting filter.
-     * @return Cold flow list of favourite motorcycles.
+     * @return Cold flow list of deleted motorcycles.
      */
     fun getLocalMotorcyclesSortedByModel(filter: MotorcyclesFilter): Flow<List<Motorcycle>> {
         return when (filter) {
-            MotorcyclesFilter.ALPHA_ASC -> localDataSource.getLocalMotorcyclesSortedByModelAsc()
-            MotorcyclesFilter.ALPHA_DESC -> localDataSource.getLocalMotorcyclesSortedByModelDesc()
+            MotorcyclesFilter.ALPHA_ASC ->
+                localDataSource.getLocalMotorcyclesSortedByModelAsc()
+            MotorcyclesFilter.ALPHA_DESC ->
+                localDataSource.getLocalMotorcyclesSortedByModelDesc()
         }
     }
 
     /**
-     * Inserts a favourite motorcycle in the local DB.
+     * Inserts or deletes a deleted motorcycle in the local DB.
      *
-     * @param motorcycle Motorcycle marked as favourite.
+     * @param motorcycle Motorcycle marked for removal.
      */
-    suspend fun saveLocalMotorcycle(motorcycle: Motorcycle) {
-        localDataSource.saveLocalMotorcycle(motorcycle)
-    }
-
-    /**
-     * Deletes a favourite motorcycle in the local DB.
-     *
-     * @param motorcycle Motorcycle marked as favourite.
-     */
-    suspend fun deleteLocalMotorcycle(motorcycle: Motorcycle) {
-        localDataSource.deleteLocalMotorcycle(motorcycle)
-    }
-
-    suspend fun deleteAllLocalMotorcycles() {
-        localDataSource.deleteAllLocalMotorcycles()
+    suspend fun updateLocalMotorcycle(motorcycle: Motorcycle) {
+        if (motorcycle.deleted) {
+            localDataSource.saveLocalMotorcycle(motorcycle)
+        } else {
+            localDataSource.deleteLocalMotorcycle(motorcycle)
+        }
     }
 }
